Удалил прослойку в виде NetworkManager, теперь запросами занимается репозиторий.
