+++ MarvelHeroes/MarvelHeroes/Application/AppDelegate.swift
@@ -20,12 +20,13 @@ final class AppDelegate: UIResponder, UIApplicationDelegate
 
 		let networkService = NetworkService()
 		let jsonDataFetcher = JSONDataFetcher(networkService: networkService)
-		let networkManager = NetworkManager(networkService: networkService, jsonDataFetcher: jsonDataFetcher)
+		let marvelItemsRepository = MarvelItemsRepository(networkService: networkService, jsonDataFetcher: jsonDataFetcher)
 
 		let modulesFactory = ModulesFactory(
 			networkService: networkService,
 			jsonDataFetcher: jsonDataFetcher,
-			networkManager: networkManager)
+			repository: marvelItemsRepository
+		)
 
 		window?.rootViewController = modulesFactory.createTabBarControllerModule()
 		window?.makeKeyAndVisible()
